requirements.txt.txt
streamlit
google-generativeai
python-pptx
fpdf